+++ CHANGELOG.md
@@ -24,6 +24,16 @@ the following form: YYYY.0M.0D.
 
 - Designed a proof of concept for MediaRatingsTracker component
 
+## [2025.03.25]
+
+### Added
+
+- Designed abstract class for MediaRatings component
+
+### Updated
+
+- Added removeAny() method to interface
+- Changed parameters for pair to String and Integer
 
 ## [2025.02.04]